split element.tsx
